Changed share name to link to match WV; moved events util to util folder
