+++ src/scripts/components/link/wv.link.shortener.js
